# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "kokodah222@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/favorite": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Get list of user favorites",
                "description": "Returns list of user favorites",
                "operationId": "getUserFavoritesList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Favorite not found"
                    }
                }
            }
        },
        "/favorite/add/{product_id}": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Store new cart product",
                "description": "Store new cart product",
                "operationId": "storeNewFavoriteProduct",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/favorite/delete/{product_id}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "cart delete product",
                "description": "Cart delete product",
                "operationId": "FavoriteDeleteProduct",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id for delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cart not found"
                    }
                }
            }
        },
        "/favorite/clear": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "favorite clear",
                "description": "Favorite clear",
                "operationId": "FavoriteClear",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cart error"
                    }
                }
            }
        },
        "/favorite/total": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "favorite total",
                "description": "Favorite total",
                "operationId": "FavoriteTotal",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Cart error"
                    }
                }
            }
        },
        "/order/make/{id}": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "create new order",
                "description": "Create new order",
                "operationId": "MakeNewOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of user",
                "description": "Returns list of user",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
...
}